helloworld/requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cachetools==5.5.0
certifi==2024.2.2
cffi==1.16.0
channels==3.0.5
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==42.0.8
daphne==3.0.2
distlib==0.3.8
dj-database-url==2.2.0
Django==5.0.6
django-cors-headers==4.3.1
django-extensions==3.2.3
django-on-heroku==1.1.2
django-sslserver==0.22
filelock==3.14.0
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
gunicorn==22.0.0
h11==0.14.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
MarkupSafe==3.0.2
msgpack==1.0.8
oauthlib==3.2.2
packaging==24.1
paypalrestsdk==1.13.3
pillow==10.3.0
pipenv==2023.12.1
platformdirs==4.2.2
proto-plus==1.25.0
protobuf==5.29.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
pyparsing==3.2.1
pytz==2024.2
redis==5.0.7
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
service-identity==24.1.0
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.3
virtualenv==20.26.2
Werkzeug==3.1.3
whitenoise==6.7.0
zope.interface==6.4.post2
